Pass calculated id to pubsub on destroy.
Use the original id calculated before deleting to pass to pubsub
instead of the id returned by the deletion operation.

The idHelper function was used to help determine the value and
a proper id may not be returned by the deletion result function.

There is no apparent reason not to use the original id for both cases.