Make educated guesses about options when specifying a blueprint route in `config/routes.js`
If you have a model User and a route config `'/user': {blueprint: 'find', limit: 5}` it will automatically set the model option to `user` for you, and set all the associations options.  You can specify an `associations` option manually with an array of strings representing the keys that you'd like to populate, e.g. `['pets', 'children']`.