Allow to set updateFiles in sails.config.i18n
This is a port of https://github.com/balderdashy/sails/blob/b49873dc0d5afa8e0f296ea48bfe1fc838364393/lib/hooks/i18n/index.js

In the stable branch, updateFiles is set to false always (you can't set it in the i18n config file). As it defaults to false, it won't break anything (unless you set it to true and left it there without working - which is unlikely).

I tested it on 0.9.16 and works as expected. I'm not sure if the "defaults: {...}" is needed in the 0.9 branch.