Fix grunt launching when it's a peer dep
There are times where someone may have grunt-cli installed as a peer
to sails. In those cases, the grunt-cli bin script isn't where we
thought it would be.

This patch uses a thin wrapper around grunt-cli, so we can exec it no
matter where it was actually installed. It just needs to be in the path
where require() can find it.