Make res.set() chainable
This fixes a bug which causes Sails to become unresponsive (or even
crash) when a message portraying an OPTIONS request for an undefined
route is emitted over the socket.io REST interface.

Steps to reproduce:

In a terminal:
$ sails new killme
$ cd killme
$ sails lift

In a browser's JS console (on localhost:1337):
> io.socket.request({method: 'options', url: '/anUndefinedRoute'});

The issue is a result of using Express' router with a constructed
response object which does not meet expectations set by Express.
For OPTIONS requests, Express' router enters a code path where send()
is called on the result of res.set() (see line 188 of lib/router/index.js
in Express v3.x). When/if set() is not chainable, we trigger an
infinite loop which brings Sails down.

The patch could have undesirable side effects if the value previously
returned from set() is used somewhere. I could not find any such
occurences in this repository. In any case, this should be considered
a quick-fix, since ideally, we should not use a third-party router at
all, to avoid similar issues (in other words, fix the TODO on line 46
in router/index.js).

The issue happens with Express v3.x, possibly not an issue in later
versions where the relevant code seems to have been subjected to
major refactorizations.