RFC / FR allow different model representations
I'm looking for a solution to selectively show/hide or transform model properties depending on the request. Similar to removing `password` before sending a user record, but dependent on whether or not the user is authorized and who that user is.

Imagine a website that is available to public visitors, but shows more information for logged in users. Now imagine the same behavior on a RESTful JSON API. When implementing such an API the blueprints for Sails are extremely helpful, so I would like to find a solution without the need to implement custom controller actions.

One possible way of expressing this could be to support different flavors of `ModelRecord.toJSON`, like `toJSONPublic`.