Update README.md
Hi, 

I added section "Books" and a link to the Sails.js in Action written by Michael R. McNeil and Irl I. Nathan for Manning Publications. We believe this book is a grat resource and all members of the community will benefit from this information. If you find this inapproprirate, please let me know where I can publish it.