Support passing configuration options to the default body parser
I'd like to be able to set config options for the default ('skipper') body parser without having to explicitly depend on it, and pass it as a body-parser function in my app.

e.g. in config/hooks.js

```javascript
module.exports.http = {
    ...
    middleware: {
        bodyParser: {
            options: {
              limit:'10mb'
            }
        }
    }
    ...
}

As it is, I am forced to pass `fn: require('skipper')` and add skipper to my dependencies, which is a bit longwinded to simply pass an option.

By defaulting the `fn` option to skipper if the bodyParser config is an object then Sails can support passing options only to the default body parser.