Expose the `loadAndRegisterControllers` method for now.
`loadAndRegisterControllers` used to be defined directly on the hook object, but was removed when it was moved to a different file.  A couple of apps in the wild are relying on it, so we'll re-introduce it for now, but it really needs to be wrapped in a public `.reload()` method.