Ensure req._sails is always set
Certain parts of the codebase - in particular the built is response functions access req._sails to gain access to log functions. However, the setting of req._sails is currently bundled with the "request" hook, which contains potentially undesirable code - in particular around parsing req.params.all(), which can have a severe performance impact in certain situations - so may be disabled.

By setting `req._sails` by default on all requests we can ensure that the built in response helpers are usable indepenently of what intermediate hooks and middleware have loaded.