Fix #3590
When using Redis adapter, inspects the `client` property and adds connection and error handlers.  Future-proofed against changes to connect-redis, and backwards compatible with cases where the user supplies an existing Redis connection in the session config as `sails.config.session.client`.

Would be nice to have tests for this, maybe using a Docker container...