Revert "Don't allow `origin: '*'` and `credentials: true` at the same time."
This reverts commit ca43e0507af79f15361789a3489013b01c8e1825.

After discussion with @mikermcneil, and in light of the fact that this was a breaking change, we decided that the previous behavior (while perhaps unexpected from the perspective of the CORS spec) was acceptable.  The default (documented) behavior if all you do is flip `allRoutes` to true in your CORS config is to leave your app wide open, which is completely acceptable (and in some cases desirable) as long as you are careful not to expose anything dangerous.  Apps can be tightened up by setting `allRoutes` to false, or changing the default `origin`, or by setting CORS properties on a per-route basis.