Remove `www` and `favicon` from the default middleware order, adding them after the express server is initialized instead.
In previous Sails versions (that used Express <4), the router was added as part of the middleware stack in sails.config.http.middleware.order, so we could just put these middleware after `router` in that list. In Express 4, the router is built in, so we have to wait until the server is fully initialized and then add the post-router middleware after.