Replace all references to `sessionConfig` with `app.config.session` for clarity.
This avoids confusion as to where the `sails.config.session` object is actually being manipulated.  It seemed for instance like we could make the session store a private var in this code, but it was actually being directly attached to the config object which is directly passed in as options to the raw Express middleware.  It's still possible we could clone the initial `sails.config.session` at the beginning and make the _whole cloned thing_ a private var -- the aim being to not expose the adapter and store to the end user -- but we'll tackle that at another time.