When disabling pubsub hook, delete it from `sails.hooks` rather than setting it to false
This is mainly to make the output nicer when inspecting the Sails object (with it set to `false`, you get a blank entry in the list of hooks), but it's also good in case someone does `_.isUndefined(sails.hooks.pubsub)` for some reason.  It's alright for the _sails.config.hooks.foo_ to be false, but `sails.hooks.foo` should either be a Hook object or `undefined`.