Handle legacy policy config, + tests
Migrating this by hand could be tedious for large config/policy.js files, and there's no chance of old/new configs colliding since the former uses a key->dictionary approach and the latter is strictly key->array.  So, we can transform the old to the new and log a warning that includes the transformed object for them to copy and paste, reducing the effort in upgrading to Sails 1.0 a bit.