Demote http hook's `defaults` function down to a lowly dictionary
It's not doing anything dynamic, so no need for it to be called as a function.  We're gonna reference the defaults when we verify that an app's configured middleware actually exists, so wanted to do this now so it doesn't inadvertently break stuff if we do it in the future.