Remove Waterline dependency, and set sails-hook-orm dep back to 1.0.9.
The direct Waterline dependency in Sails is no longer used (Waterline is a direct dependency of sails-hook-orm instead).

The sails-hook-orm dep was bumped so that upcoming changes to Waterline (v0.11.7 prerelease) could be tested, but now that Waterline 0.11.7 has been published, it will be picked up automatically by new installs of sails-hook-orm 1.0.9.