Better logic around getting/setting _middlewareType when binding functions
Setting _middlewareType in the `options` argument to `.bind()` is only supported for cases where it's not practical to set it directly on the function itself (like when the function is being created inline in the call to `.bind()` ).  So once it's been read, we remove it from the options so that it doesn't get merged in to the `req.options` dictionary that eventually winds up in the request handler.