Improve the way CORS and CSRF hooks operate
Instead of looping through sails.config.routes (explicit routes), bind handlers for `router:bind` and `route:typeUnknown` so that they can gather up lists of routes to operate on, and then bind a _single_ /* route that loops through the list, sees if the current request matches a route that needs CORS or CSRF applied, and does its thing.  This allows routes that were defined in hooks to have CSRF and CORS options applied directly.  Also improves the "set-preflight-config" function to not have to do `pathToRegexp` at request time.

Committing this to a branch right now b/c we're so close to v1 release, and because there's a workaround for hook routes, which is to declare identical routes in `config/routes.js` with the CORS / CSRF config you want and _no_ target.