handle E_UNIQUE in replace blueprint action
The only time "replace" would encounter a uniqueness error is if there was concurrent logic running somewhere that was messing around with the database at the same time  (it could never occur due to duplicates in the array of new child ids because of [this](https://github.com/balderdashy/waterline/blob/7b3e5f3229a9caef4bdb6a5e3cdb4868d0717141/lib/waterline/utils/query/private/normalize-pk-value-or-values.js#L89-L91)).  Since we don't currently use transactions in blueprints (and because, in some cases e.g. xD/A or adapters that don't support `join`, we'll never be able to use transactions), we handle this with a 500 error for now (because in the database could be in an a intermediate state, and this endpoint doesn't attempt to rectify that).  In the future, in the case where this blueprint action was able to use a transaction, it should respond with a 409 instead (because any changes we made would have been rolled back automatically due to the failure).