NODE_ENV only has special meaning if it === 'production'
This replaces a "isUndefined or `==='development'`" check with `!=='production'`

If your NODE_ENV is set to "foobar", or "test", or "staging", or "prod", or "ANYTHING" other than `production`, we take that to mean that you're in development mode.  NODE_ENV is used strictly to check for the "effective production-ness".  For more flexibility, use `sails.config.environment`.  For example, you might want a particular staging environment to run with NODE_ENV=production, but a different staging server emulating a developer box to run _without_ NODE_ENV set.  Similarly, you  might want to run your tests with _and_ without NODE_ENV=production.