Allow dots in module filenames.
This allows config files like `/config/foo.bar.js` to be read in, instead of being ignored.  It's likely that the previous regex was intended to capture everything up to the final `.` as the filename, it just wasn't working quite as expected.  Note that for modules that derive their identity from the filename, it's up to the hook that loads the module to deal with dots.  For example, sails-hook-orm will fail to lift with a helpful error if you try to name a file `api/models/Foo.bar.js`.