Update ExampleHelper.md to use `const`
The sails docs use `var` in most example code. This is an outdated practice as the newer alternatives of `let` and `const` offer more granular control over variable scoping in a way that is preferable 95+% of the time. The use of `var` in the docs gives new users the impression that sails is dated. Updating the docs to use `const` or occasionally `let` in the examples will make sails more appealing.

I'm opening this PR as a one-off for a single page in the hopes of getting feedback on whether the sails team agrees that this change is appropriate before I more holistically propose changes to a broader range of sails docs to update them similarly.